Add modal for important messages, bugfixes

+++ SensorplatformApp/db/create-tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE HeartRateSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 transmitted boolean NOT NULL,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
@@ -9,7 +9,7 @@ rrintervals VARCHAR(100) NOT NULL
 )
 
 CREATE TABLE PulseOximeterSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 spo2 INTEGER NOT NULL,
@@ -17,7 +17,7 @@ pulserate INTEGER NOT NULL
 )
 
 CREATE TABLE CC2650TemperatureSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 objecttemperature FLOAT NOT NULL,
@@ -25,7 +25,7 @@ dietemperature FLOAT NOT NULL
 )
 
 CREATE TABLE CC2650PressureSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 temperature FLOAT NOT NULL,
@@ -33,7 +33,7 @@ pressure FLOAT NOT NULL
 )
 
 CREATE TABLE CC2650HumiditySample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 temperature FLOAT NOT NULL,
@@ -41,14 +41,14 @@ humidity FLOAT NOT NULL
 )
 
 CREATE TABLE CC2650AmbientlightSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 ambientlight FLOAT NOT NULL
 )
 
 CREATE TABLE CC2650MovementSample (
-id INTEGER IDENTITY,
+id BIGINT IDENTITY,
 timestamp VARCHAR(30) NOT NULL,
 device VARCHAR(17) NOT NULL,
 rotationX FLOAT,
